% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CanadaC19_demo}
\alias{CanadaC19_demo}
\title{Demo of CanadaC19 cases data.}
\format{
A tibble object
}
\source{
\href{https://github.com/ishaberry/Covid19Canada}{website}
}
\usage{
CanadaC19_demo
}
\description{
daily summary of Canadian Coronavirus (COVID-19) cases.
}
\details{
This package includes tidy data of Canadian COVID-19 cases.

To get the latest data, use \code{\link{refresh_CanadaC19_cases}}
}
\examples{

data(CanadaC19_demo)

#Show us something cool

}
\keyword{COVID19}
\keyword{Canada}
\keyword{coronavirus}
\keyword{datasets}
