% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_CanadaC19.R
\name{get_info_CanadaC19}
\alias{get_info_CanadaC19}
\title{Get information about the datasets provided by YOURPACKAGENAME}
\usage{
get_info_CanadaC19()
}
\value{
a tibble of information about the datasets in this package
}
\description{
Returns information about the datasets in this package for covid19R harvesting
}
\section{Functions}{
\itemize{
\item \code{get_info_CanadaC19}: Basic information about the dataset.
}}

\examples{
\dontrun{

# get the dataset info from this package
get_info_CanadaC19()
}

}
