% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_CanadaC19.R
\name{refresh_CanadaC19_cases}
\alias{refresh_CanadaC19_cases}
\title{Canadian COVID-19 Data}
\source{
\href{https://github.com/ishaberry/Covid19Canada}{Covid19Canada}

\href{https://github.com/Covid19R/documentation}{covid19R documentation}
}
\usage{
refresh_CanadaC19_cases(verbose = TRUE)
}
\arguments{
\item{verbose}{Should messages be displayed?}
}
\value{
A tibble object
\itemize{
\item date - The date in YYYY-MM-DD form
\item location - The name of the location as provided by the data source. XXXXX
\item location_type - The type of location using the covid19R controlled vocabulary. SAY MORE ABOUT THIS
\item location_code - A standardized location code using a national or international standard. ADD MORE
\item location_code_type The type of standardized location code being used according to the covid19R controlled vocabulary. Here we use \code{XXXX}
\item data_type - the type of data in that given row using the covid19R controlled vocabulary. Includes WHAT DATA TYPES ARE HERE?
\item value - number of cases of each data type
}
}
\description{
Updates the Canadian COVID-19 cases.
}
\details{
This package scrapes Canadian reported cases from a repository that collects publically available data on cases across Canada.
}
\examples{
\dontrun{
refresh_CanadaC19_cases()
}
}
